REFACTOR: Change display grid with flex:
The collection section is intended to show various products of the same family.
In desktop version, to avoid a repetitive layout, the idea is to flip each row
in order to break the compocisiton.

To avoid messing with the html, the idea was to use a CSS rule to flip the layout.
I only know how to do it in Flex display.
That's the argument for the change