UPDATE: Replace new model for test. Add new assets for future experiment:
Bah! is not an experiment, but I want to document the process
to export an obj from Sketchup, and assemble it in Blender
in order to export it with Sparkart Toolkit (this framework
would be a good fit if I add the service web + social media)

But because it does a mess with the real axis (basically
presented rotated (Y axis function as X and viceversa)
I'll have to add a separated page with the same documentation
but not for SparkarToolset. Maybe an index like with the 'Laboratory
' page